+++ src/spas/home/components/ScoreCard/index.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class ScoreCard extends Component {
+  render() {
+    return <div>{this.props.children}</div>;
+  }
+}
+
+export default ScoreCard;
