+++ src/spas/home/index.js
@@ -1,40 +1,18 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import Card from './components/Card';
-import Monte from './components/Monte';
 import './styles.scss';
 import './assets/home.font';
-
-function shuffleArray(array) {
-  let i = array.length - 1;
-  for (; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    const temp = array[i];
-    array[i] = array[j];
-    array[j] = temp;
-  }
-  return array;
-}
+import CardTable from './components/CardTable';
+import ScoreCard from './components/ScoreCard';
 
 class HomeSPA extends Component {
   constructor() {
     super();
 
-    this.shuffler = setInterval(() => {
-      this.handleShuffle();
-    }, 500);
-
     this.state = {
       score: 0,
-      winningScore: 10,
-      cards: ['one', 'two', 'three']
+      winningScore: 10
     };
-  }
-
-  componentDidMount() {
-    setTimeout(() => {
-      clearInterval(this.shuffler);
-    }, 3000);
   }
 
   setWinningScore() {
@@ -46,29 +24,15 @@ class HomeSPA extends Component {
     return tally;
   }
 
-  handleShuffle() {
-    this.setState(prevState => {
-      const newOrder = shuffleArray(prevState.cards);
-
-      return {
-        cards: newOrder
-      };
-    });
-  }
-
   render() {
     return (
       <div className="container">
         <div className="content">
-          <div className="card-container">
-            <Card id={this.state.cards[0]}>
-              <Monte />
-            </Card>
-            <Card id={this.state.cards[1]} />
-            <Card id={this.state.cards[2]} />
-          </div>
+          <CardTable />
         </div>
-        <div className="toolbar">{this.setWinningScore()}</div>
+        <div className="toolbar">
+          <ScoreCard>{this.setWinningScore()}</ScoreCard>
+        </div>
       </div>
     );
   }
